Add Event Statistics

+++ app/Http/Controllers/eventController.php
@@ -17,8 +17,13 @@ class eventController extends Controller
     /**
      * @OA\Get(
      *     path="/api/event",
-     *     summary="Get all events records",
+     *     summary="Lấy tất cả các sự kiện",
      *     tags={"Event"},
+     *      description="
+     *      - Endpoint trả về thông tin của tất cả các sự kiện
+     *      - Role được sử dụng là role của tất cả 
+     *      - Trả về thông tin của tất cả các sự kiện đã diễn ra
+     * ",
      *     @OA\Response(
      *         response=200,
      *         description="Successful operation",
@@ -91,6 +96,10 @@ class eventController extends Controller
      *     summary="Tìm kiếm sự kiện theo tên ",
      *     tags={"Event"},
      *     operationId="Tìm kiếm sự kiện",
+     *  description="
+     * - Request cần nhập vào là tên sự kiện
+     * - Tên sự kiện chỉ cần nhập gần giống, không nhất thiết phải giống hẳn
+     * - Endpoit trả ra là những sự kiện có tên dạng vậy",
      *     @OA\RequestBody(
      *         required=true,
      *         @OA\JsonContent(
@@ -169,8 +178,17 @@ class eventController extends Controller
      * @OA\Post(
      *     path="/api/event",
      *     tags={"Event"},
-     *     summary="Store a new event record",
-     *     description="Thêm mới bản ghi với dữ liệu được cung cấp",
+     *     summary="Thêm mới bản ghi với dữ liệu được cung cấp",
+     *     description="
+     * - Endpoint trả về bản ghi mới được thêm vào 
+     * -Role đước sử dụng là nhân viên, quản lí
+     * -name là tên sự kiện 
+     * -location là nơi tổ chức sự kiện 
+     * -contact là liên lạc bằng số điện thoại
+     * -user_id là id của user tổ chức sự kiện này
+     * -start_time là thời gian bắt đầu sự kiện
+     * -end_time là thời gian kết thúc sự kiện
+     * ",
      *     operationId="storeEvent",
      *     @OA\RequestBody(
      *         required=true,
@@ -282,7 +300,7 @@ class eventController extends Controller
         }
         return response([
             "status" => "error",
-            "message" => "Only Employees and managers are allowed to add new records",
+            "message" => "Chỉ nhân viên và quán lí mới có thể tạo mới sự kiện",
             "statusCode" => Response::HTTP_CONFLICT
         ], Response::HTTP_CONFLICT);
     }
@@ -292,8 +310,10 @@ class eventController extends Controller
      *      path="/api/event/{id}",
      *      operationId="getEventsById",
      *      tags={"Event"},
-     *      summary="Get events by ID",
-     *      description="Lấy dữ liệu sự kiện từ một id cho trước",
+     *      summary="Lấy dữ liệu sự kiện từ một id cho trước",
+     *      description="
+     * -Endpoint này lấy ra 1 bản ghi của sự kiện
+     * -id là id của event mà mình cần tìm kiếm",
      *      @OA\Parameter(
      *          name="id",
      *          description="ID sự kiện ",
@@ -360,18 +380,143 @@ class eventController extends Controller
             ], Response::HTTP_NOT_FOUND);
         }
     }
-
+/**
+     * @OA\Post(
+     *     path="/api/eventStatistics",
+     *     tags={"Event"},
+     *     summary="Thống kê sự kiện ",
+     *     description="
+     * - Endpoint trả về các bản ghi sự kiện
+     * -Role đước sử dụng quản lí, nhân viên
+     * -name là tên sự kiện 
+     * -location là nơi tổ chức sự kiện 
+     * -contact là liên lạc bằng số điện thoại
+     * -user_id là id của user tổ chức sự kiện này
+     * -start_time là thời gian bắt đầu sự kiện
+     * -end_time là thời gian kết thúc sự kiện
+     * -attendances_count là số sinh viên tham gia
+     *  - attendances là thông tin của các sinh viên tham gia
+     *  - feedback là thông tin của các feedback của sinh viên
+     * - start_time có thể rỗng 
+     * - end_time có thể rỗng
+     * - Nhưng nếu nhập start_time thì bắt buộc phải nhập end_time
+     * - Nếu không nhập cả start_time và end_time thì sẽ là thống kê của tuần hiện tại
+     * ",
+     *     operationId="eventStatistics",
+     *     @OA\RequestBody(
+     *         required=true,
+     *         @OA\JsonContent(
+     *             @OA\Property(property="start_time", type="string",format="date-time", example="2023-11-23 11:20:22"),
+     *             @OA\Property(property="end_time", type="string",format="date-time", example="2023-11-23 11:20:22"),
+     *         )
+     *     ),
+     *     @OA\Response(
+     *         response=200,
+     *         description="Dữ liệ trả về thành công",
+     *         @OA\JsonContent(
+     *             @OA\Property(property="status", type="string", example="success"),
+     *             @OA\Property(property="message", type="string", example="Tạo mới bản ghi thành công"),
+     *             @OA\Property(property="statusCode", type="int", example=200),
+     *             @OA\Property(property="metadata", type="array",
+     *                  @OA\Items(type="object",
+     *                           @OA\Property(property="name", type="string", example="Event Name"),
+     *                           @OA\Property(property="location", type="string", example="Ha Noi"),
+     *                           @OA\Property(property="contact", type="string", example="0986567467"),
+     *                           @OA\Property(property="user_id", type="integer", example=2),
+     *                           @OA\Property(property="start_time", type="string",format="date-time", example="2023-11-23 11:20:22"),
+     *                           @OA\Property(property="end_time", type="string",format="date-time", example="2023-11-23 11:20:22"),
+     *                           @OA\Property(property="attendances_count", type="integer", example=0),
+     *                           @OA\Property(property="feedback", type="array", @OA\Items(
+     *                              type="object",
+     *                      @OA\Property(property="id", type="integer", example="1"),
+     *                     @OA\Property(property="content", type="string", example="Phucla DepZai"),
+     *                     @OA\Property(property="user_id", type="integer", example="1"),
+     *                     @OA\Property(property="event_id", type="integer", example="2"),
+     *                     @OA\Property(property="created_at", type="string", format="date-time", example="2023-11-28 17:02:29"),
+     *                     @OA\Property(property="updated_at", type="string", format="date-time", example="2023-11-28 17:02:29"),
+     *                          )),
+     *                         @OA\Property(property="attendances", type="array", @OA\Items(
+     *                              type="object",
+     *                     @OA\Property(property="id", type="interger", example=1),
+     *                     @OA\Property(property="user_id", type="integer", example=1),
+     *                     @OA\Property(property="event_id", type="integer", example=2),
+     *                          )),
+     *                  )
+     *             )
+     *         )
+     *     ),
+     *     @OA\Response(
+     *         response=422,
+     *         description="Sai validate",
+     *         @OA\JsonContent(
+     *             @OA\Property(property="status", type="string", example="error"),
+     *             @OA\Property(property="message", type="string", example="Sai validate"),
+     *             @OA\Property(property="statusCode", type="int", example=422),
+     *         )
+     *     ),
+     *     @OA\Response(
+     *         response=500,
+     *         description="Lỗi hệ thống",
+     *         @OA\JsonContent(
+     *             @OA\Property(property="status", type="string", example="error"),
+     *             @OA\Property(property="message", type="string", example="Lỗi hệ thống"),
+     *             @OA\Property(property="statusCode", type="int", example=500),
+     *         )
+     *     )
+     * )
+     */
     public function eventStatistics(Request $request)
     {
+        $logUser = auth()->user()->role;
+        if($logUser == 0){
+            return response([
+                "status" => "error",
+                "message" => 'Sinh viên không thể xem thống kê',
+                'statusCode' => Response::HTTP_INTERNAL_SERVER_ERROR
+            ], Response::HTTP_INTERNAL_SERVER_ERROR);
+        }
         //Nếu không có request thì là mặc định tuần hiện tại
-        if ($request == null) {
+        if ($request->start_time == "" && $request->end_time == "") {
             $today = Carbon::now();
 
             // Lấy thông tin về tuần và năm
             $weekNumber = $today->weekOfYear;
             $year = $today->year;
-            return $weekNumber;
+            $dayOfWeekNumber = $today->dayOfWeek;
+
+            //Lấy ngày đầu tiên, cuối cùng của tuần đó
+            $firstDayOfWeekNumber = $today->copy()->addDays(-$dayOfWeekNumber+1);
+            $lastDayOfWeekNumber = $today->copy()->addDays($dayOfWeekNumber+1);
+            $eventInWeek = event::where('end_time','>=',$firstDayOfWeekNumber)
+                                                ->where('start_time','<=',$lastDayOfWeekNumber)
+                                                ->with('feedback')
+                                                ->withCount('attendances')
+                                                ->with('attendances')
+                                                ->get();
+            return $eventInWeek;
         }
+        $validator = Validator::make($request->all(),[
+            'start_time'=>'required',
+            'end_time'=>'required', 'after:start_time'
+        ],[
+            'start_time.required'=>'Ngày bắt đầu phải có',
+            'end_time.required'=>'Ngày kết thúc phải có',
+            'start_time.after'=>'Ngày kết thúc phải lớn hơn ngày bắt đầu'
+        ]);
+        if ($validator->fails()) {
+            return response([
+                "status" => "error",
+                "message" => $validator->errors(),
+                'statusCode' => Response::HTTP_INTERNAL_SERVER_ERROR
+            ], Response::HTTP_INTERNAL_SERVER_ERROR);
+        }
+        $eventInStatistic = event::where('end_time','>=',$request->start_time)
+        ->where('start_time','<=',$request->end_time)
+        ->with('feedback')
+        ->withCount('attendances')
+        ->with('attendances')
+        ->get();
+        return $eventInStatistic;
     }
 
     /**
@@ -380,8 +525,17 @@ class eventController extends Controller
      *     path="/api/event/{id}",
      *     operationId="updateEvent",
      *     tags={"Event"},
-     *     summary="Update Event",
-     *     description="Sửa dữ liệu của một sự kiện ",
+     *     summary="Sửa dữ liệu của một sự kiện ",
+     *     description="
+     * -Endpoint trả về dữ liệu bản ghi mới được sửa đổi
+     * -Role quy định là role nhân viên, quản lí
+     * -name là tên sự kiện 
+     * -location là nơi tổ chức sự kiện 
+     * -contact là liên lạc bằng số điện thoại
+     * -user_id là id của user tổ chức sự kiện này
+     * -start_time là thời gian bắt đầu sự kiện
+     * -end_time là thời gian kết thúc sự kiện
+     * ",
      *     @OA\Parameter(
      *         name="id",
      *         description="ID của một sự kiện",
@@ -504,7 +658,7 @@ class eventController extends Controller
         }
         return response([
             "status" => "error",
-            "message" => "Only Employees and managers are allowed to edit records",
+            "message" => "Chỉ nhân viên và quản lí mới có quyền sửa bản ghi",
             "statusCode" => Response::HTTP_CONFLICT
         ], Response::HTTP_CONFLICT);
     }
@@ -514,6 +668,12 @@ class eventController extends Controller
      *     path="/api/event/{id}",
      *     summary="Xóa một bản ghi",
      *     tags={"Event"},
+     * description="
+     *          - Endpoint này sẽ xóa 1 sự kiện 
+     *          - Role được sử dụng là role Quản lí
+     *          - Xóa thành công sẽ trả lại data là của các sự kiện còn lại
+     *          - id là id của event cần xóa
+     *          ",
      *     @OA\Parameter(
      *         name="events",
      *         in="path",
@@ -564,8 +724,18 @@ class eventController extends Controller
                     'statusCode' => Response::HTTP_NOT_FOUND
                 ], Response::HTTP_NOT_FOUND);
             }
+            $logUserRole = auth()->user()->role;
+            if($logUserRole != 2){
+                return response()->json([
+                    'message' => 'Không phải quản lí thì sẽ không có quyền xóa',
+                    'status' => 'error',
+                    'statusCode' => Response::HTTP_CONFLICT
+                ], Response::HTTP_CONFLICT);
+            }
             $event->delete();
+            $restOfEvents = event::all();
             return response()->json([
+                'metadata'=> $restOfEvents,
                 'message' => 'Xóa bản ghi thành công',
                 'status' => 'success',
                 'statusCode' => Response::HTTP_OK
